remove useEffect for updating detectorIds
